Deprecate old init.py, add new init.py
